requirements.txt
pygame>=2.0.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
numpy>=1.21.0 